+++ tests/legacy/test_tron.py
@@ -137,7 +137,7 @@ async def test_tron_no_models():
 
 async def test_mid_out_of_range(tron_client, tron_server):
 
-    tron_client.send_command("actor", "command", mid=(2 ** 32 + 2))
+    tron_client.send_command("actor", "command", mid=(2**32 + 2))
 
     assert 2 in tron_client.running_commands